+++ mdcx/crawlers/__init__.py
@@ -50,7 +50,7 @@ from .javbus import JavbusCrawler
 from .javday import JavdayCrawler
 from .javdb import JavdbCrawler
 from .javdb_api import JavdbApiCrawler
-from .javdb_app import JavdbAPICrawler
+from .javdb_app import JavdbAppCrawler
 from .javlibrary import JavlibraryCrawler
 from .kin8 import Kin8Crawler
 from .libredmm import LibredmmCrawler
@@ -72,7 +72,7 @@ from .xcity import XcityCrawler
 register_crawler(DmmCrawler)
 register_crawler(JavdbCrawler)
 register_crawler(JavdbApiCrawler)
-register_crawler(JavdbAPICrawler)
+register_crawler(JavdbAppCrawler)
 register_crawler(dmm_api.DmmApiCrawler)
 register_crawler(AvbaseCrawler)
 register_crawler(missav.MissavCrawler)